added startBed stub to tmak, added goroutine to start bed and create session

+++ database/structs.go
@@ -36,5 +36,5 @@ type Session struct {
 	Id          int `db:"autoInc"`
 	Bed_num     int
 	Customer_id int
-	Time_stamp  int
+	Time_stamp  int64
 }
